host/requirements.txt
pyserial
vgamepad
pystray
pillow
